[ui] Extract combine-units.tsx from search-bulk-action-button.tsx
