+++ app/src/app/search/components/search-bulk-action-button.tsx
@@ -1,10 +1,11 @@
+"use client";
 import { Button } from "@/components/ui/button";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
 } from "@/components/ui/popover";
-import { Combine, Settings, Trash } from "lucide-react";
+import { Settings, Trash } from "lucide-react";
 import { Command } from "cmdk";
 import {
   CommandEmpty,
@@ -14,49 +15,11 @@ import {
   CommandList,
 } from "@/components/ui/command";
 import * as React from "react";
-import { cn } from "@/lib/utils";
-import { useRouter } from "next/navigation";
 import { useCartContext } from "@/app/search/cart-provider";
+import CombineUnits from "@/app/search/components/bulk-actions/combine-units";
 
 export default function SearchBulkActionButton() {
-  const router = useRouter();
   const { selected, clearSelected } = useCartContext();
-
-  const isEligibleForCombination =
-    selected.length === 2 &&
-    selected.find((unit) => unit.unit_type === "enterprise") &&
-    selected.find((unit) => unit.unit_type === "legal_unit");
-
-  const setPrimaryLegalUnitForEnterprise = async () => {
-    const legalUnit = selected.find((unit) => unit.unit_type === "legal_unit");
-    const enterprise = selected.find((unit) => unit.unit_type === "enterprise");
-
-    if (!legalUnit || !enterprise) {
-      console.error(
-        "failed to set primary legal unit for enterprise due to missing legal unit or enterprise"
-      );
-      return;
-    }
-
-    try {
-      const response = await fetch(
-        `/api/legal-units/${legalUnit.unit_id}/primary`,
-        {
-          method: "POST",
-          body: JSON.stringify(enterprise),
-        }
-      );
-
-      if (!response.ok) {
-        console.error("failed to set primary legal unit for enterprise");
-        return;
-      }
-
-      router.push(`/enterprises/${enterprise.unit_id}`);
-    } catch (e) {
-      console.error("failed to set primary legal unit for enterprise", e);
-    }
-  };
 
   return (
     <Popover>
@@ -80,22 +43,7 @@ export default function SearchBulkActionButton() {
           <CommandList>
             <CommandEmpty>No command found.</CommandEmpty>
             <CommandGroup>
-              <CommandItem
-                disabled={!isEligibleForCombination}
-                onSelect={setPrimaryLegalUnitForEnterprise}
-                className={cn(
-                  "flex-col items-start space-y-1",
-                  !isEligibleForCombination && "opacity-50"
-                )}
-              >
-                <div className="flex items-center space-x-2">
-                  <Combine className="h-4 w-4" />
-                  <span>Combine units</span>
-                </div>
-                <span className="text-xs">
-                  One Legal Unit and one Enterprise
-                </span>
-              </CommandItem>
+              <CombineUnits />
             </CommandGroup>
             <CommandGroup>
               <CommandItem onSelect={clearSelected} className="space-x-2">
